Remove ArrayTransformer and move related methods to more appropriate places
