+++ jvm-agent/src/main/org/jetbrains/lincheck/jvm/agent/LincheckClassVisitor.kt
@@ -143,10 +143,6 @@ internal class LincheckClassVisitor(
         chain.addTransformer { adapter, mv ->
             ObjectCreationTransformer(fileName, className, methodName, desc, access, methodInfo, context, adapter, mv)
         }
-        // ======== Reflection and ArrayCopy ========
-        chain.addTransformer { adapter, mv ->
-            ArrayTransformer(fileName, className, methodName, desc, access, methodInfo, context, adapter, mv, instrumentationMode == EXPERIMENTAL_MODEL_CHECKING)
-        }
 
         // ======== Method Calls ========
         chain.addTransformer { adapter, mv ->
