+++ jvm-agent/src/main/org/jetbrains/lincheck/jvm/agent/TransformationProfile.kt
@@ -34,7 +34,6 @@ class TransformationConfiguration(
     var trackStaticFieldWrites: Boolean = false,
     var trackArrayElementReads: Boolean = false,
     var trackArrayElementWrites: Boolean = false,
-    var trackArrayCopy: Boolean = false,
 
     var trackMethodCalls: Boolean = false,
     var interceptMethodCallResults: Boolean = false,
@@ -128,19 +127,18 @@ class TransformationConfiguration(
 
     val trackSharedMemoryAccesses: Boolean
         get() = trackRegularFieldReads  || trackStaticFieldReads  || trackArrayElementReads ||
-                trackRegularFieldWrites || trackStaticFieldWrites || trackArrayElementWrites ||
-                trackArrayCopy
+                trackRegularFieldWrites || trackStaticFieldWrites || trackArrayElementWrites
+
 
     var trackAllSharedMemoryAccesses: Boolean
         get() =
             trackAllFieldsReads && trackArrayElementReads &&
-            trackAllFieldsWrites && trackArrayElementWrites && trackArrayCopy
+            trackAllFieldsWrites && trackArrayElementWrites
         set(value) {
             trackAllFieldsReads = value
             trackArrayElementReads = value
             trackAllFieldsWrites = value
             trackArrayElementWrites = value
-            trackArrayCopy = value
         }
 
     val trackThreadsOperations: Boolean
@@ -196,7 +194,6 @@ internal fun TransformationConfiguration.shouldApplyVisitor(visitorClass: Class<
         ThrowTransformer::class.java -> trackThrows
         CatchBlockStartTransformer::class.java -> trackCatchBlocks
 
-        ArrayTransformer::class.java -> trackArrayCopy
 
         // the configuration does not govern other types of transformers,
         // so they should be applied by default
@@ -576,7 +573,6 @@ object ExperimentalModelCheckingTransformationProfile : TransformationProfile {
             trackLocalVariableWrites = true
             trackArrayElementReads = true
             trackArrayElementWrites = true
-            trackArrayCopy = true
 
             interceptReadResults = true
         }
