+++ jvm-agent/src/main/org/jetbrains/lincheck/jvm/agent/transformers/ObjectCreationTransformer.kt
@@ -19,6 +19,9 @@ import org.objectweb.asm.Opcodes.*
 import org.objectweb.asm.Type
 import org.objectweb.asm.commons.GeneratorAdapter
 import org.objectweb.asm.commons.InstructionAdapter.OBJECT_TYPE
+import org.objectweb.asm.commons.Method
+import java.lang.StringBuilder
+import kotlin.reflect.KFunction
 
 /**
  * [ObjectCreationTransformer] tracks creation of new objects,
@@ -79,6 +82,12 @@ internal class ObjectCreationTransformer(
     private var uninitializedObjects = 0
 
     override fun visitMethodInsn(opcode: Int, owner: String, name: String, desc: String, itf: Boolean) = adapter.run {
+
+        if(opcode == INVOKESTATIC && owner == "java/lang/reflect/Array" && name == "newInstance") {
+            visitInvokeArrayNewInstance(opcode, owner, name, desc, itf)
+            return
+        }
+
         if (name != "<init>") {
             super.visitMethodInsn(opcode, owner, name, desc, itf)
             return
@@ -349,5 +358,82 @@ internal class ObjectCreationTransformer(
         //   to skip instrumenting allocations of immutable types,
         //   to avoid the runtime overhead when immutable-value tracking is disabled.
         bootstrapMethodOwner == "java/lang/invoke/StringConcatFactory"
+
+
+    private fun visitInvokeArrayNewInstance(opcode: Int, owner: String, name: String, descriptor: String, isInterface: Boolean) = adapter.run {
+        // TODO: should also call beforeNewObjectCreation?
+        // STACK: elementClass, length
+        invokeIfInAnalyzedCode(
+            original = {
+                visitMethodInsn(opcode, owner, name, descriptor, isInterface)
+            },
+            instrumented = {
+                // STACK: elementClass, length
+                swap()
+                // STACK: length, elementClass
+                dup()
+                // STACK: length, elementClass, elementClass
+                val elementClass = newLocal(OBJECT_TYPE).also { storeLocal(it) }
+                // STACK: length, elementClass
+                swap()
+                // STACK: elementClass, length
+                visitMethodInsn(opcode, owner, name, descriptor, isInterface)
+                // STACK: array
+                dup()
+                // STACK: array, array
+                invokeStatic(ThreadDescriptor::getCurrentThreadDescriptor)
+                // STACK: array, array, descriptor
+                swap()
+                loadLocal(elementClass)
+                // STACK: array, array, descriptor, elementClass
+                getElementTypeNameFromClass()
+                // STACK: array, descriptor, array, elementName
+                invokeStatic(Injections::afterObjectConstructor)
+                // STACK: array
+            }
+        )
+    }
+
+    /**
+     * Converts the class of the array elements into the canonical java string represetnation such that
+     * it can be used by [Injections::afterObjectConstructor]
+     * NOTE: Probably could be written as a regular method.
+     *
+     * Stack beore: ..., elementClass : Class<*>
+     * Stack after: ..., elementClassName : String
+     */
+    private fun GeneratorAdapter.getElementTypeNameFromClass() {
+        val getTypeFun: (Class<*>) -> Type = Type::getType
+        val sbType = Type.getType(java.lang.StringBuilder::class.java)
+        val typeType = Type.getType(Type::class.java)
+        val sbConstructor = Method.getMethod(java.lang.StringBuilder::class.java.getDeclaredConstructor())
+        val appendMethod = Method.getMethod(java.lang.StringBuilder::class.java.getDeclaredMethod("append", String::class.java))
+        val toStringMethod = Method.getMethod(StringBuilder::class.java.getDeclaredMethod("toString"))
+        val getClassNameMethod = Method.getMethod(Type::class.java.getDeclaredMethod("getClassName"))
+
+        // elementClass
+        invokeStatic(getTypeFun as KFunction<*>)
+        // elementType
+        // NOTE: we use StringBuilder for Java8 compatibility
+        newInstance(sbType)
+        dup()
+        // elementType, stringBuilder
+        invokeConstructor(sbType, sbConstructor)
+        // elementType, stringBuilder
+        swap()
+        // stringBuilder, elementType
+        // NOTE: getClassName handles the annoying part where arrays elements are handled differently,
+        //   than all the other types
+        invokeVirtual(typeType, getClassNameMethod)
+        // stringBuilder, elementString
+        invokeVirtual(sbType, appendMethod)
+        // stringBuilder
+        push("[]")
+        // stringBuilder, "[]"
+        invokeVirtual(sbType, appendMethod)
+        // stringBuilder,
+        invokeVirtual(sbType, toStringMethod)
+        // name
+    }
 
 }
